+++ R/DR2S-package.R
@@ -7,18 +7,11 @@
 #' @docType package
 #' @name DR2S
 #' @useDynLib DR2S
-#' @importFrom methods is
 #' @import assertthat
-#' @import data.table
-#' @import foreach
-#' @import ggplot2
 #' @import futile.logger
-#' @import ipdDb
-#' @import rtracklayer
-#' @import Rsamtools
-#' @importFrom cowplot save_plot plot_grid
-#' @importFrom iterators iter nextElem
-#' @importFrom Rcpp compileAttributes
-#' @importFrom ggseqlogo geom_logo theme_logo
-#' @importFrom rlang .data set_names
+#' @import ggplot2
+#' @importFrom data.table :=
+#' @importFrom foreach foreach %do%
+#' @importFrom rlang .data
+#' @importFrom methods is
 NULL